// Engine/Source/Programs/Horde/HordeDashboard/src/backend/Api.ts
export enum Priority {

	// Lowest priority
	Lowest = "Lowest",

	// Below normal priority
	BelowNormal = "BelowNormal",

	// Normal priority
	Normal = "Normal",

	// Above normal priority
	AboveNormal = "AboveNormal",

	// High priority
	High = "High",

	// Highest priority
	Highest = "Highest"
}
// ...
export enum JobStepBatchState {

	// Waiting for dependencies of at least one jobstep to complete
	Waiting = "Waiting",

	// Getting ready to execute
	Starting = "Starting",

	// Ready to execute
	Ready = "Ready",

	// Currently running
	Running = "Running",

	// Getting ready to execute
	Stopping = "Stopping",

	// All steps have finished executing
	Complete = "Complete"
}

/**Error code for a batch not being executed */
export enum JobStepBatchError {

	/** No error */
	None = "None",

	/** The stream for this job is unknown */
	UnknownStream = "UnknownStream",

	/** The given agent type for this batch was not valid for this stream */
	UnknownAgentType = "UnknownAgentType",

	/** The pool id referenced by the agent type was not found */
	UnknownPool = "UnknownPool",

	/** There are no agents in the given pool currently online */
	NoAgentsInPool = "NoAgentsInPool",

	/** There are no agents in this pool that are onlinbe */
	NoAgentsOnline = "NoAgentsOnline",

	/** Unknown workspace referenced by the agent type */
	UnknownWorkspace = "UnknownWorkspace",

	/** Cancelled */
	Cancelled = "Cancelled",

	/** Lost connection with the agent machine */
	LostConnection = "LostConnection",

	/** Lease terminated prematurely */
	Incomplete = "Incomplete",

	/** An error ocurred while executing the lease. Cannot be retried. */
	ExecutionError = "ExecutionError",

	/** The change that the job is running against is invalid. */
	UnknownShelf = "UnknownShelf",

	/** No longer needed */
	NoLongerNeeded = "NoLongerNeeded",

	/** Sync Failed */
	SyncingFailed = "SyncingFailed"

}
// ...
export type AgentData = GetAgentResponse
export type LeaseData = GetAgentLeaseResponse
export type SessionData = GetAgentSessionResponse
// ...
export enum LeaseState {

	Unspecified = "Unspecified",

	Pending = "Pending",

	Active = "Active",

	Completed = "Completed",

	Cancelled = "Cancelled"

}


export enum LeaseOutcome {

	/** Default value. */
	Unspecified = "Unspecified",

	/** The lease was executed successfully */
	Success = "Success",

	/** The lease was not executed succesfully, but cannot be run again. */
	Failed = "Failed",

	/** The lease was cancelled by request */
	Cancelled = "Cancelled"

}


/**Response for queries to find a particular lease within an agent */
export type GetAgentLeaseResponse = {

	/**Identifier for the lease */
	id: string;

	/** parent lease id */
	parentId?: string;

	/**Name of the lease */
	name?: string;

	/**Time at which the lease started (UTC) */
	startTime: Date | string;

	/**Time at which the lease ended (UTC) */
	finishTime?: Date | string;

	/**Whether this lease has started executing on the agent yet */
	executing: boolean;

	/**Details for this lease */
	details?: { [key: string]: string };

	/**agentId of lease */
	agentId?: string;

	/**What type of lease this is (parsed from Details array) */
	type: string;

	/**BatchId of this lease (parsed from Details array) */
	batchId: string;

	/**JobId of this lease (parsed from Details array) */
	jobId: string;

	/**logId of this lease (parsed from Details array) */
	logId: string;

	/**Batch data for this lease, queried from backend */
	batch?: GetBatchResponse;

	/** Outcome of the lease */
	outcome?: LeaseOutcome;

	state?: LeaseState;

}

/**Information about an agent session */
export type GetAgentSessionResponse = {

	/**Unique id for this session */
	id: string;

	/**Start time for this session */
	startTime: Date | string;

	/**Finishing time for this session */
	finishTime?: Date | string;

	/**Capabilities of this agent */
	capabilities?: GetAgentCapabilitiesResponse;

	/**Version of the software running during this session */
	version?: string;

}

export type GetAgentWorkspaceResponse = {

	/**Server and port */
	serverAndPort?: string;

	/**username */
	userName?: string;

	/**password */
	password?: string;

	/**identifier */
	identifier: string;

	/**stream */
	stream: string;

	/**view */
	view?: string[];

	/**incremental */
	bIncremental: boolean;
}

/**Information about an agent */
export type GetAgentResponse = {

	/**The agent's unique ID */
	id: string;

	/**Friendly name of the agent */
	name: string;

	/**Session ID */
	sessionId: string;

	/**Session expiry time */
	sessionExpiresAt: Date | string;

	/**Status of the agent */
	online: boolean;

	/**Whether the agent is currently enabled */
	enabled: boolean;

	/**Whether the agent is ephemeral */
	ephemeral: boolean;

	/**comment for the agent */
	comment?: string;

	/**The current client version */
	version?: string;

	/**Per-agent override for the desired client version */
	forceVersion?: string;

	/**Pools for this agent */
	pools?: string[];

	/**Status for this agent */
	status?: string;

	/**Capabilities of this agent */
	capabilities?: GetAgentCapabilitiesResponse;

	/**Array of active leases. */
	leases?: GetAgentLeaseResponse[];

	/**Per-object permissions */
	acl?: GetAclResponse;

	/**Last update time of the agent */
	updateTime: Date | string;

	/**Whether this agent is deleted */
	deleted: boolean;

	/** Whether agent is pending conform */
	pendingConform: boolean;

	/** Whether a full conform job is pending */
	pendingFullConform: boolean;

	/** Conform attempt count */
	conformAttemptCount?: number;

	/** Last conform time */
	lastConformTime?: Date | string;

	/** Next conform attempt time */
	nextConformTime?: Date | string;

	/** The reason for the last shutdown */
	lastShutdownReason: string;

	/** Whether a shutdown is pending */
	pendingShutdown: boolean;

	/** agent workspaces */
	workspaces?: GetAgentWorkspaceResponse[];

}
// ...
export type GetReportResponse = {

	// Report placement
	placement: ReportPlacement;

	/** Name of the report */
	name: string;

	/** The artifact id */
	// artifactId?: string;

	/** The report markdown content */
	content?: string;

	/** The job or step message */
	message?: string;

	/** The job or step severity */
	severity?: string;

	/** Color to display for message severity */
	color?: string;
}

export type GetJobArtifactResponse = {
	/// Identifier for this artifact
	id: string;
	/// Name of the artifact
	name: string;
	/// Artifact type
	type: string;
	/// Description to display for the artifact on the dashboard
	description?: string;
	/// Step producing the artifact
	stepId: string;
	/// Keys for the artifact	
	keys: string[];
	/// Metadata for the artifact
	metadata: string[];
}

/**Information about a job */
export type GetJobResponse = {

	/**Unique Id for the job */
	id: string;

	/**Unique id of the stream containing this job */
	streamId: string;

	/**Name of the job */
	name: string;

	/** The changelist number to build */
	change?: number;

	/** The preflight changelist number */
	preflightChange?: number;

	/** Description of the preflight */
	preflightDescription?: string;

	/** Commit hash for branch-based preflights */
	preflightCommitId?: string;

	/** The template type */
	templateId?: string;

	/** Hash of the actual template data */
	templateHash?: string;

	/** Hash of the graph for this job */
	// graphHash?: string;

	/** The user that started this job */
	startedByUserInfo?: GetThinUserInfoResponse;

	/** The user that started this job */
	abortedByUserInfo?: GetThinUserInfoResponse;

	/** The reason the job was canceled */
	cancellationReason?: string;

	/** Whether job was created by a bisect task */
	startedByBisectTaskId?: string;

	/** The roles to impersonate when executing this job */
	roles: string[];

	/** Priority of the job */
	priority: Priority;

	/** Whether the change will automatically be submitted or not */
	autoSubmit?: boolean;

	/** The submitted changelist number */
	autoSubmitChange?: number;

	/**  Message produced by trying to auto-submit the change */
	autoSubmitMessage?: string;

	/**Time that the job was created */
	createTime: Date | string;

	/** The global job state */
	state: JobState;

	/** Array of jobstep batches */
	batches?: GetBatchResponse[];

	/**  List of labels */
	labels?: GetLabelStateResponse[];

	/** The default label, containing the state of all steps that are otherwise not matched. */
	defaultLabel?: GetLabelStateResponse;

	/** List of reports */
	reports?: GetReportResponse[];

	/**  Parameters for the job */
	arguments: string[];

	/** Parameters for the job */
	parameters: Record<string, string>;

	/**The last update time for this job*/
	updateTime: Date | string;

	/** Whether to update issues based on the outcome of this job */
	updateIssues?: boolean;

	artifacts?: GetJobArtifactResponse[];

	/**  Custom permissions for this object */
	acl?: GetAclResponse;

	/** Additional arguments for the job, job dialog v2 */
	additionalArguments?: string[];

	/** Custom list of targets for the job, will take precedence over all argument and parameter targets */
	targets?: string[];

	/**  Whether the current user is allowed to update this job */
	canUpdate?: boolean;

	/// The parent job id, if any
	parentJobId?: string;

	/// The parent job step id, if any
	parentJobStepId?: string;

}
// ...
export type JobStepId = string;
// ...
export type GetBatchResponse = {

	/**Unique id for this batch */
	id: string;

	/**The unique log file id */
	logId?: string;

	/**Index of the group being executed */
	groupIdx: number;

	/**The state of this batch */
	state: JobStepBatchState;

	/**Error code for this batch */
	error: JobStepBatchError;

	/**Steps within this run */
	steps: GetStepResponse[];

	/**The agent assigned to execute this group */
	agentId?: string;

	/// The agent type
	agentType: string;

	/** The USD rate of an agent hour */
	agentRate?: number;

	/**The agent session holding this lease */
	sessionId?: string;

	/**The lease that's executing this group */
	leaseId?: string;

	/**The priority of this batch */
	weightedPriority: number;

	/**Time at which the group became ready (UTC) */
	readyTime?: Date | string;

	/**Time at which the group started (UTC). */
	startTime?: Date | string;

	/**Time at which the group finished (UTC) */
	finishTime?: Date | string;

	/** Human-readable names of lease outcome reasons (only populated when batch has an error) */
	leaseOutcomeReasons?: string[];
}
// ...
export type GetAgentTelemetrySampleResponse = {
	time: Date;
	userCpu: number;
	idleCpu: number;
	systemCpu: number;
	freeRam: number;
	usedRam: number;
	totalRam: number;
	workingFreeDisk: number;
	workingTotalDisk: number;
	systemFreeDisk: number;
	systemTotalDisk: number;
}
// ...
export type GetAgentCapabilitiesResponse = {

	/**Information about the devices required for this node to run */
	devices?: GetDeviceCapabilitiesResponse[];

	/**Global agent properties for this node */
	properties?: string[];

}
// ...
export type GetAclResponse = {
	/**Entries to replace the existing ACL */
	entries: GetAclEntryResponse[];

	/**Whether to inherit permissions from the parent entity */
	inheritPermissions: boolean;
}
// ...
export enum LabelState {

	/** label is not currently being built (no required nodes are present)*/
	Unspecified = "Unspecified",

	/** Steps are still running */
	Running = "Running",

	/** All steps are complete */
	Complete = "Complete"
}

/// Outcome of an aggregate
export enum LabelOutcome {

	/** Aggregate is not currently being built */
	Unspecified = "Unspecified",

	/** A step dependency failed */
	Failure = "Failure",

	/** A dependency finished with warnings */
	Warnings = "Warnings",

	/** Successful */
	Success = "Success"
}


/** State of the job */
export enum JobState {

	/** Waiting for resources*/
	Waiting = "Waiting",

	/** Currently running one or more steps*/
	Running = "Running",

	/** All steps have completed */
	Complete = "Complete"
}
// ...
export type GetLabelStateResponse = {

	// Name to show for this label on the dashboard
	dashboardName?: string;

	// Category to show this label in on the dashboard
	dashboardCategory?: string;

	// Name to show for this label in UGS
	ugsName?: string;

	// Project to display this label for in UGS
	ugsProject?: string;

	// Steps to include in the status of this label
	steps: JobStepId[];

	/**State of the label */
	state?: LabelState;

	/**Outcome of the label */
	outcome?: LabelOutcome;
}
// ...
export type GetThinUserInfoResponse = {

	/**  Id of the user */
	id: string;

	/** Name of the user */
	name: string;

	/**  The user's email address */
	email: string;

}